Implmement #45-getRoom and partial #45-createRoom
Room class skeleton has been created. getRoom currently searches database
for a single room belonging to an admin. If more than one room for an
admin is found, an error is thrown.

createRoom has been commented out, and needs to be fully fleshed out.

Note: AdminPage.js has been updated with an example of how to use such
a function.