Feature backend join room hook (#63)
* Created the tasklist class and task class

* Created Admin class

* Add tasklist creation ability on admin page

* Comment code on creating tasklist

* skeleton for Room

* Implmement #45-getRoom and partial #45-createRoom

Room class skeleton has been created. getRoom currently searches database
for a single room belonging to an admin. If more than one room for an
admin is found, an error is thrown.

createRoom has been commented out, and needs to be fully fleshed out.

Note: AdminPage.js has been updated with an example of how to use such
a function.

* implement room to firestore and add roomErrors

implement the createRoom, getRoom, getOrCreateroom with firestore

Note: we can use realtime database, and we just have to change out
how queries happen.

TODO: implement
- deleteRoom, joinRoom, leaveRoom

* Add tasklist creation ability on admin page

* Comment code on creating tasklist

* Create tasklist object on AdminPage

* Handle adding a new task to the task list

* description for updateTask

* update description

* modify Admin class

* modify Admin class

* modify Tasklist

* modify Tasklist

* basic task list

-not hooked up to backend
-ugly, needs beautification
-remember to discard changes to AdminPage before merging to main since there are changes to AdminPage in roomcode branch

* joinRoom handle (#51)

classmethod to allow players to join a room defined with "roomCode".

The method currently adds playerId to a players list, however this
is a dirty placeholder and needs to be refactored to add `Player`s
to the list, when the object exists.

* CSS Changes for AdminPage and TaskList

Made CSS adjustments to TaskList in AdminPage

* update comments

* Resolved bugs

A Little bit of visual clean up

* Room Code Generation Added

Diverts to another page once start a room clicked

* renamed folders

* Cleaned up admin page

* Fixed directory names

* refactor Room and database rooms doc id to use roomCode

Room objects in the database are now querried using the roomCode.

roomCode is the key for the database.

* joinRoom handle (#51)

classmethod to allow players to join a room defined with "roomCode".

The method currently adds playerId to a players list, however this
is a dirty placeholder and needs to be refactored to add `Player`s
to the list, when the object exists.

update comments

* update joinRoom

---------

Co-authored-by: Hongkun Li <hongkun@Hongkuns-MacBook-Air.local>
Co-authored-by: NamrataAcharya02 <nnachary@purdue.edu>
Co-authored-by: Hongkun Li <hongkun@Hongkuns-Air.lan>
Co-authored-by: Hongkun Li <hongkun523@gmail.com>
Co-authored-by: bluepaw59 <96216128+bluepaw59@users.noreply.github.com>
Co-authored-by: pinlunix <pheonlu@gmail.com>